:recycle: create a `network` module
The details of initializing and running the network stack are now,
mostly, in the new `network` module.
